+++ libdap2/daputil.c
@@ -4,6 +4,7 @@
  *********************************************************************/
 
 #include "config.h"
+#include <stddef.h>
 
 #ifdef HAVE_SYS_TIME_H
 #include <sys/time.h>
@@ -312,7 +313,7 @@ collectocpath(OClink conn, OCddsnode node, NClist* path)
 char*
 makeocpathstring(OClink conn, OCddsnode node, const char* sep)
 {
-    int i,len,first;
+    size_t i,len,first;
     char* result;
     char* name;
     OCtype octype;
@@ -353,7 +354,7 @@ makeocpathstring(OClink conn, OCddsnode node, const char* sep)
 char*
 makepathstring(NClist* path, const char* separator, int flags)
 {
-    int i,len,first;
+    size_t i,len,first;
     NCbytes* pathname = NULL;
     char* result;
     CDFnode* node;
@@ -412,7 +413,7 @@ clonenodenamepath(CDFnode* node, NClist* path, int withdataset)
 char*
 simplepathstring(NClist* names,  char* separator)
 {
-    int i;
+    size_t i;
     size_t len;
     char* result;
     if(nclistlength(names) == 0) return nulldup("");
